+++ src/main/resources/static/js/delete_turno.js
@@ -1,6 +1,10 @@
 function deleteBy(id)
 {
 
+        let aler = confirm("Esta por eliminar un turno, si aun desea hacerlo haz click en confirmar");
+
+         if (aler == true){
+
           const url = '/turnos/'+ id;
           const settings = {
               method: 'DELETE'
@@ -10,5 +14,11 @@ function deleteBy(id)
 
           let row_id = "#tr_" + id;
           document.querySelector(row_id).remove();
+
+          }else{
+
+                      alert("Excelente decision!");
+
+                     }
 
 }
